+++ src/components/features/ArticleModifyModal/index.tsx
@@ -9,11 +9,11 @@ import InputUI from '../../ui/InputUI';
 import SelectUI from '../../ui/SelectUI';
 import CheckboxUI from '../../ui/CheckboxUI';
 import { TitleInputSection, ButtonContainer } from './style';
-import { useUpdateInfoRequest, useGetInfoRequest } from '../../../api-requests/Info';
-import { useUpdateNoticeRequest, useGetNoticeRequest } from '../../../api-requests/Notice';
-import { useGetAllCategoryRequest } from '../../../api-requests/InfoCategory';
-import { useGetMyInfoRequest, useGetMyNoticeRequest } from '../../../api-requests/MyPage';
-import { useUploadImageRequest } from '../../../api-requests/ImageUpload';
+import { useUpdateInfoRequest, useGetInfoRequest } from '../../../api-hooks/Info';
+import { useUpdateNoticeRequest, useGetNoticeRequest } from '../../../api-hooks/Notice';
+import { useGetAllCategoryRequest } from '../../../api-hooks/InfoCategory';
+import { useGetMyInfoRequest, useGetMyNoticeRequest } from '../../../api-hooks/MyPage';
+import { useUploadImageRequest } from '../../../api-hooks/ImageUpload';
 import toast from 'react-hot-toast';
 import { useSelectedArticleAtom } from '../../../store/articleAtom';
 import { useCurrentLocationAtom } from '../../../store/currentLocationAtom';
