+++ src/components/ui/CategoryUI/AddMode/index.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { Tr, Td } from '@chakra-ui/react';
 import InputUI from '../../InputUI';
 import { DefaultButton, CancelButton } from '../../ButtonUI';
-import useCreateCategoryRequest, { useGetCategoryRequest } from '../../../../api-requests/InfoCategory';
+import useCreateCategoryRequest, { useGetCategoryRequest } from '../../../../api-hooks/InfoCategory';
 import toast from 'react-hot-toast';
 
 interface AddModeProps {
