+++ src/pages/InfoPage/InfoManagePage/index.tsx
@@ -17,8 +17,8 @@ import TextViewer from '../../../components/features/TextViewer';
 import Pagination from '../../../components/features/Pagination';
 import { SimpleGrid, Image, Flex } from '@chakra-ui/react';
 import { RepeatIcon } from '@chakra-ui/icons';
-import { useGetInfoRequest } from '../../../api-requests/Info';
-import { useGetAllCategoryRequest } from '../../../api-requests/InfoCategory';
+import { useGetInfoRequest } from '../../../api-hooks/Info';
+import { useGetAllCategoryRequest } from '../../../api-hooks/InfoCategory';
 import toast from 'react-hot-toast';
 import { changeDateFormat } from '../../../utils/changeDateFormat';
 import { useSelectedArticleAtom } from '../../../store/articleAtom';
